lower method routing to be a bool
holding a target for each method is a waste of memory, it's better to
optimize for the most common case.

Part of the original reason here was to eventually reduce the number of
eql(name) checks and then require matches to be unique, but even still,
there are better ways to go about that.